Improve performance for analyzing images
