+++ src/lib/extractStreamMeta.ts
@@ -2,7 +2,9 @@ import { spawn } from 'duplex-child-process';
 import { Stream } from 'stream';
 
 export default async (stream: Stream): Promise<any> => new Promise((resolve, reject) => {
-  const proc = spawn('magick', ['-', 'json:']);
+  // [1x1+0+0] only analyzes a small portion of the image instead of the full
+  // increasing the speed drastically
+  const proc = spawn('magick', ['-[1x1+0+0]', 'json:']);
   stream.pipe(proc);
   const buffer = [];
   proc.on('data', (chunk) => {
